API: add support for buffer exports.
Add interfaces for low-level buffer exports to suport interop with external APIs like EGL or OpenCL
(OCL). Theory of operations:

- vaAcquireBufferHandle(): locks buffer for external API usage. This
  is a synchronization point. Upon successful return, useful buffer
  details are returned in a somewhat implementation-specific way to
  support interop with external APIs.

- vaReleaseBufferHandle(): unlocks buffer after usage from external
  API, and deallocates any resources that were needed.

This is a low-level API that replaces any former API specific to EGL, DRM or OpenCL interop. External
API specific extensions may be provided in the future, in view to simplifying the end usages. The
relevant implementations would be using those low-level APIs to achieve the same result, with less
code.

Signed-off-by: Gwenole Beauchesne <gwenole.beauchesne@intel.com>